+++ keyboards/electronlab/klor/keymaps/electronlab/rules.mk
@@ -1,12 +1,12 @@
 OLED_ENABLE = yes
 OLED_DRIVER = ssd1306
 ENCODER_ENABLE = yes
-#ENCODER_MAP_ENABLE = yes
+ENCODER_MAP_ENABLE = yes
 EXTRAKEY_ENABLE = yes
 AUDIO_ENABLE = yes
 MUSIC_ENABLE = yes
 DYNAMIC_MACRO_ENABLE = yes
-#COMBO_ENABLE = yes
+COMBO_ENABLE = yes
 KEY_OVERRIDE_ENABLE = no
 TRI_LAYER_ENABLE = yes
 SPLIT_KEYBOARD = yes
